Extracted Timeouts from AkkaBakerConfig
